+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Added selectedDays prop
 
 ## 3.2.0
 * Fixed an issue #21: because of default first day of week, default locale is changed to en_GB